DA#229: Modified: Refactor handlers code from extension.ts
